:art: Feat: 도메인 구조로 변경
-도메인 구조로 변경

Ref: #17
