+++ src/main/java/com/greenity/server/activity/dto/ApplicationDTO.java
@@ -1,0 +1,14 @@
+package com.greenity.server.activity.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ApplicationDTO {
+    private long activityId;
+    private String activityTitle;
+    private String isAccepted;
+}
