--- src/main/java/com/greenity/server/mapper/ActivityRoundMapper.java
@@ -1,32 +1,0 @@
-package com.greenity.server.mapper;
-
-import com.greenity.server.dto.ActivityRoundDTO;
-import com.greenity.server.model.ActivityRound;
-
-public class ActivityRoundMapper {
-
-    public static ActivityRoundDTO toDTO(ActivityRound activityRound) {
-        return new ActivityRoundDTO(
-                activityRound.getActivityDate(),
-                activityRound.getActivityTime(),
-                activityRound.getActivityNum(),
-                activityRound.getLocation()
-        );
-    }
-
-    public static ActivityRound toEntity(ActivityRoundDTO dto) {
-        ActivityRound activityRound = new ActivityRound();
-        activityRound.setActivityDate(dto.getActivityDate());
-        activityRound.setActivityTime(dto.getActivityTime());
-        activityRound.setActivityNum(dto.getActivityNum());
-        activityRound.setLocation(dto.getLocation());
-        return activityRound;
-    }
-
-    public static void updateEntity(ActivityRound activityRound, ActivityRoundDTO dto) {
-        activityRound.setActivityDate(dto.getActivityDate());
-        activityRound.setActivityTime(dto.getActivityTime());
-        activityRound.setLocation(dto.getLocation());
-        // 회차 번호는 보통 수정하지 않으므로 제외. 필요 시 추가 가능.
-    }
-}
