--- src/main/java/com/greenity/server/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.greenity.server.repository;
-
-import com.greenity.server.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-}
