+++ src/main/java/com/greenity/server/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.greenity.server.user.repository;
+
+import com.greenity.server.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+}
